minor: add license and notice (#4)
This pull request makes two main changes: it updates the project's
license from MIT to Apache 2.0, and it adds a new `NOTICE` file to
comply with the requirements of the Apache license. Additionally, there
is a minor update to the markdown linting configuration.

**License and Legal Compliance:**

* Replaced the contents of the `LICENSE` file with the full text of the
Apache License, Version 2.0, updating the project's licensing terms from
MIT to Apache 2.0.
* Added a new `NOTICE` file to provide attribution and comply with
Apache 2.0 license requirements.

**Tooling and Configuration:**

* Updated `.markdownlint.json` to disable rule `MD010` (hard tabs),
allowing for more flexibility in markdown formatting.